api/requirements.txt
fastapi
uvicorn
psycopg[binary]
python-jose
passlib[argon2]
argon2-cffi